requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
atomic-agents==1.1.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.3.0
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==8.0.1
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
helium==5.1.1
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
instructor==1.7.9
Jinja2==3.1.6
jiter==0.8.2
linkify-it-py==2.0.3
lxml==5.3.2
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
mcp==1.1.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
multidict==6.4.3
openai==1.75.0
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.7
primp==0.15.0
propcache==0.3.1
pydantic==2.10.3
pydantic_core==2.27.1
pyfiglet==1.0.2
Pygments==2.19.1
PySocks==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
readability-lxml==0.8.1
requests==2.32.3
rich==13.9.4
selenium==4.31.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
smolagents==1.14.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sse-starlette==2.2.1
starlette==0.46.2
tenacity==9.1.2
terminal-img==0.0.3
textual==0.89.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.2
typing_extensions==4.13.2
uc-micro-py==1.0.3
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.0
